+++ test/SemanticVersionTest.groovy
@@ -1,8 +1,8 @@
+import static org.junit.Assert.assertEquals
+
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
-import static org.junit.Assert.*
-
 
 @RunWith(Parameterized.class)
 class SemanticVersionTest {
@@ -16,10 +16,14 @@ class SemanticVersionTest {
 
     @Test
     void sortsCorrectly() {
-        List<SemanticVersion> unsorted = SORTED.clone().collect({v -> new SemanticVersion(v)})
-        Collections.shuffle(unsorted)
-        def result = unsorted.sort().collect({sv -> sv.version})
-        assertEquals(SORTED,result)
+        List<SemanticVersion> versions = SORTED.clone().collect { v -> new SemanticVersion(v) }
+        // Randomize
+        Collections.shuffle(versions)
+        // Then sort
+        versions.sort()
+
+        def result = versions.collect { sv -> sv.version }
+        assertEquals(SORTED, result)
     }
 
     static SORTED = [
@@ -30,22 +34,22 @@ class SemanticVersionTest {
             '0.1.0.2',
             '0.1.0.10',
             '0.1.0.23',
-            '0.2', 
-            '0.2.0.3.1', 
+            '0.2',
+            '0.2.0.3.1',
             '0.2.0.4',
-            '1.0RC1', 
-            '1.0RC2', 
-            '1.0', 
-            '1.0.1.2', 
-            '1.0.2', 
+            '1.0RC1',
+            '1.0RC2',
+            '1.0',
+            '1.0.1.2',
+            '1.0.2',
             '1.2.2.3',
-            '1.2.3', 
-            '1.5.2_04', 
-            '1.5.2_05', 
-            '1.5.2_10', 
+            '1.2.3',
+            '1.5.2_04',
+            '1.5.2_05',
+            '1.5.2_10',
             '1.6.0_01',
-            '2.0', 
-            '2.0.0_02', 
+            '2.0',
+            '2.0.0_02',
             '3.1'
     ]
 }