+++ test/TerraformPluginVersion11Test.groovy
@@ -1,14 +1,9 @@
-import org.junit.*
+import static org.mockito.Mockito.spy
+import static org.mockito.Mockito.verify;
+
+import org.junit.Test
 import org.junit.runner.RunWith
 import de.bechte.junit.runners.context.HierarchicalContextRunner
-
-import static org.hamcrest.Matchers.*
-import static org.junit.Assert.*
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.spy
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Matchers.*
 
 @RunWith(HierarchicalContextRunner.class)
 class TerraformPluginVersion11Test {