+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+# v5.4
+
+* [Issue #87](https://github.com/manheim/terraform-pipeline/issues/87) Add travis CI support to the project
+* [Issue #186](https://github.com/manheim/terraform-pipeline/issues/186) Add convenience method to detect the version of terraform
+* [Issue #185](https://github.com/manheim/terraform-pipeline/issues/185) Fix defect: filename parameter is not used in TerraformPlugin.readFile
+
 # v5.3
 
 * [Issue #178](https://github.com/manheim/terraform-pipeline/issues/178) Fix broken link to source code in AwssumePlugin  docs